Split the connect-oauth route into shared, forms, detail, and sections.
Move loader, query/storage helpers, and presentational cards out of the
2192-line route so it drops under the client-route file-size ratchet
without changing connect, callback, or setup behavior.
