+++ tools/file-size-ratchet.json
@@ -12,7 +12,6 @@
 		"packages/worker/client/routes/admin-platform-integrations.tsx",
 		"packages/worker/client/routes/admin-users.tsx",
 		"packages/worker/client/routes/community-detail.tsx",
-		"packages/worker/client/routes/connect-oauth.tsx",
 		"packages/worker/client/routes/home.tsx",
 		"packages/worker/client/routes/login.tsx",
 		"packages/worker/client/routes/onboarding-mcp-client-tabs.tsx",
